Update ALLNETWORKS_qaqc.py
I added the argparse library and the argument parsing in the main function. I think it might be a good idea to start implementing this in our functions, it might help in the long run with user and function support. Please let me know if this is OK with you, it is a rather simple approach: for this function, I have added `--network`  and `--verbose` flags, buy you can notice that verbose is not being used at the moment.